+++ packages/bridge-controller/src/utils/metrics/types.ts
@@ -6,6 +6,7 @@ import type {
   MetaMetricsSwapsEventSource,
   MetricsActionType,
   MetricsSwapType,
+  PollingStatus,
 } from './constants';
 import type { SortOrder, StatusTypes } from '../../types';
 
@@ -213,7 +214,7 @@ export type RequiredEventContextFromClient = {
   [UnifiedSwapBridgeEventName.StatusValidationFailed]: {
     failures: string[];
   };
-  [UnifiedSwapBridgeEventName.MaxPollingReached]: TradeData &
+  [UnifiedSwapBridgeEventName.PollingStatusUpdated]: TradeData &
     Pick<QuoteFetchData, 'price_impact'> &
     Omit<RequestMetadata, 'security_warnings'> &
     Pick<
@@ -224,20 +225,8 @@ export type RequiredEventContextFromClient = {
       | 'chain_id_destination'
     > & {
       action_type: MetricsActionType;
-      polling_attempts: number;
-    };
-  [UnifiedSwapBridgeEventName.PollingManuallyRestarted]: TradeData &
-    Pick<QuoteFetchData, 'price_impact'> &
-    Omit<RequestMetadata, 'security_warnings'> &
-    Pick<
-      RequestParams,
-      | 'token_symbol_source'
-      | 'token_symbol_destination'
-      | 'chain_id_source'
-      | 'chain_id_destination'
-    > & {
-      action_type: MetricsActionType;
-      polling_attempts: number;
+      polling_status: PollingStatus;
+      retry_attempts: number;
     };
 };
 
@@ -295,8 +284,7 @@ export type EventPropertiesFromControllerState = {
   [UnifiedSwapBridgeEventName.StatusValidationFailed]: RequestParams & {
     refresh_count: number;
   };
-  [UnifiedSwapBridgeEventName.MaxPollingReached]: null;
-  [UnifiedSwapBridgeEventName.PollingManuallyRestarted]: null;
+  [UnifiedSwapBridgeEventName.PollingStatusUpdated]: null;
 };
 
 /**